Extract admin stats into own module.
