+++ perma_web/perma/templates/includes/upper_right_menu.html
@@ -46,7 +46,7 @@
             <li><h3 class="dropdown-header">Admin Only</h3>
               <ul class="dropdown-inner-list">
                 <li><a href="{% url 'admin:index' %}">Admin console</a></li>
-                <li><a href="{% url 'user_management_stats' %}">Site stats</a></li>
+                <li><a href="{% url 'admin_stats' %}">Site stats</a></li>
               </ul>
             </li>
           <li class="divider"></li>
