+++ perma_web/perma/tests/test_views_user_management.py
@@ -2638,8 +2638,8 @@ class UserManagementViewsTestCase(PermaTestCase):
 
     def test_admin_stats(self):
         self.log_in_user(self.admin_user)
-        self.get('user_management_stats', reverse_kwargs={'args':['days']})
-        self.get('user_management_stats', reverse_kwargs={'args':['celery']})
-        self.get('user_management_stats', reverse_kwargs={'args':['random']})
-        self.get('user_management_stats', reverse_kwargs={'args':['emails']})
-        self.get('user_management_stats', reverse_kwargs={'args':['job_queue']})
+        self.get('admin_stats', reverse_kwargs={'args':['days']})
+        self.get('admin_stats', reverse_kwargs={'args':['celery']})
+        self.get('admin_stats', reverse_kwargs={'args':['random']})
+        self.get('admin_stats', reverse_kwargs={'args':['emails']})
+        self.get('admin_stats', reverse_kwargs={'args':['job_queue']})
